+++ Modules/QuestFrame/QuestFrameModule.lua
@@ -643,6 +643,8 @@ do
             local addedQuests = {}
             local displayMapIDs = DataModule:GetMapIDsToGetQuestsFrom(mapID)
 
+            local searchBoxText = QuestScrollFrame.SearchBox:GetText():lower()
+
             for mID in pairs(displayMapIDs) do
                 local taskInfo = C_TaskQuest.GetQuestsForPlayerByMapID(mID)
 
@@ -653,7 +655,7 @@ do
                                 local isFiltered = DataModule:IsQuestFiltered(info, mapID)
                                 if not isFiltered then
                                     if addedQuests[info.questId] == nil then
-                                        local button = QuestFrameModule:QuestLog_AddQuestButton(info)
+                                        local button = QuestFrameModule:QuestLog_AddQuestButton(info, searchBoxText)
 
                                         if button ~= nil then
                                             table.insert(usedButtons, button)
@@ -671,6 +673,9 @@ do
                 -- In the situation where the normal quest log is empty, but we have world quests.
                 -- We shouldn't show the empty quest log text.
                 QuestScrollFrame.EmptyText:Hide()
+
+                -- We need to also make sure the "No search results" text is hidden.
+                QuestScrollFrame.NoSearchResultsText:Hide()
             else
                 QuestFrameModule:HideWorldQuestsHeader()
                 return
@@ -698,7 +703,7 @@ do
         QuestScrollFrame.Contents:Layout()
     end
 
-    function QuestFrameModule:QuestLog_AddQuestButton(questInfo)
+    function QuestFrameModule:QuestLog_AddQuestButton(questInfo, searchBoxText)
         local questID = questInfo.questId
         local title, factionID, _ = C_TaskQuest.GetQuestInfoByQuestID(questID)
         local questTagInfo = C_QuestLog.GetQuestTagInfo(questID)
@@ -706,6 +711,10 @@ do
         C_TaskQuest.RequestPreloadRewardData(questID)
 
         if (questTagInfo == nil) then
+            return nil
+        end
+
+        if searchBoxText ~= "" and not title:lower():find(searchBoxText, 1, true) then
             return nil
         end